requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
unstructured
langchain-text-splitters
langchain-google-genai
tokenizers>=0.19.1
huggingface-hub<1.0.0,>=0.33.4
huggingface
transformers
scikit-learn
pandas==2.0.0
torch==2.0.0
speechrecognition
PyAudio
pyttsx3